mhi_integration_engine/extractors/MSreviewpro/requirements.txt
functions-framework==3.0.0
pandas==2.1.4
pyarrow==16.1.0
gcsfs==2024.6.1
Jinja2==3.1.2
google-cloud-pubsub==2.22.0
google-cloud-secret-manager==2.19.0